[FDF][ResumeTask] Support system resume
This change adds a resumetask that resumes the parent first and then
spawns resumetasks for children. The resumetask of the parent will not
complete until all the children resume. The device state transitions
from kSuspended->kResuming->kResumed->kActive, during a resume task.
Composite device resume will be handled in the next change
Test: Added tests to coordinator-test. fx run-test coordinator-tests.

Change-Id: Ia8dd0f51b26c40371ce758f8e80d2202eb8eb106