[devcoord] Remove vfs_exit timeout
There was an arbitrary 60 seconds timeout while waiting for
fshost to shutdown. For certain devices this is not enough
time to flush content to storage. In particular during a
system upgrade critical blobs might be lost resulting on a
non-functional system afterwards.

With this change we remove the timeout, so we wait until
the storage stack is done.

BUG: 38103

TEST: Existing test suffice, in particular
reboot_tests expect the following log message:
  i.WaitForLogMessage("devcoordinator: Successfully waited for VFS exit completion")

Change-Id: I762bd293052799df8bf48c3cb853b3fdd05df392