[FDF] Handle multiple bind requests correctly
When a binddriver request is issued on a device on devfs connection,
the transaction is cached. A separate fidl request issued from devhost to
devcoordinator, takes care of loading the driver and calling the Driver's bind hook.
The original cached fidl txn is completed when the bind hook returns.
Prior to this change, when there are multiple bind requests on a device
multiple fidl txns are cached but only one results in calling the bind
hook of the driver. Others silently fail and the cached transactions are
not completed.

This change properly handles the multiple requests case by returning
ZX_ERR_ALREADY_BOUND from devcoordinator, if the device is already bound.
In this case, the devhost will not queue up the fidl txn because, we can
reply about the error right away.

With this change we no longer need the workaround for fvm driver:
https://fuchsia.googlesource.com/fuchsia/+/master/zircon/system/core/devmgr/devhost/rpc-server.cc#523

Test: Added success/failure tests in devhost-test. runtests -t
devhost-test-test. while /system/test/fs/blobfs-integration-test; do true; done

Change-Id: I55e2b46454f878b73df84af53bb988a6e62e7143