[netsvc] Add --all-features command line flag to netsvc
In the fullness of time, we would like to remove netsvc entirely from
configs above bringup. However, because of various dependencies this
isn't practical today, so for now, add a limited mode to netsvc.

By default, when netsvc is launched it will started in this limited
mode. When in limited mode, netsvc will service only NB_QUERY over UDP
as well as ICMP as usual. It will also not attempt to broadcast kernel
logs.

In order to get the normal feature set from netsvc, the --all-features
command line option must now be specified. --all-features is currently
passed and set by the bringup config, which all other product
configurations inherit, so in practice when launched netsvc will
continue to work as normal.

The default limited netsvc mode supports the use of netaddr and netls
(used broadly by "fx"). However, loglistener (which is required for "fx
log"), sending files by TFTP, using netcp, or using netruncmd are all
disabled.

Adds integration tests to confirm that the listed tools either do or
don't work as desired. The test for netruncmd is currently disabled
(pending a followup CL that will remove the GN
zircon_enable_netsvc_debugging_features flag and key it on this new
netsvc feature flag).

Bug: 37043
Change-Id: I320091b9619a9b7d22bd54090d39c5859ad7df93