[FDF] Add SetPerformanceState fidl op and driver hook
Drivers can publish various performance states for devices, to conserve
power. For example, for an EMMC device, a fully performant state can be
ultra high speed, and a low performant state can be operating at lower
speed. For a CPU device, when thermal limits are
reached, we would want to put the cpu at a lower operating point to
reduce the heat generation. The SetPerformanceState fidl op allows the
system wide power manager to modify the performant state of a device.
If a set_performance_state hook is not implemented, the fidl op returns
ZX_ERR_NOT_SUPPORTED.

This change adds the fidl op and the driver hook,
adds 2 test drivers, one implementing and one without and tests the
results. The device is supposed to remember what device performant state
it is currently in.

Test: added 2 test drivers, one implementing changeperformance hook and one
without. runtests -t ddk-power-test.

Bug: 39906

Change-Id: I662569112f86b0f0018c3f53485388e88ea4e6fa