[devcoord] add 20 seconds to appmgr startup timeout
We've seen this take just a fair bit longer in infra on QEMU than we're
giving here.  Let's bump the timeout for now as a band-aid; the
longer-term solution lies in having component_manager launch appmgr
instead and have all the directory opens be pipelined, and that's
already being worked on.

Bug: 41163
Change-Id: I0f2fbfa962863b92d35aa574e707b83397c5bdd9