[fdf] Wire up init hook on devhost side.
Now all devices will have an InitTask created for them. If the device
does not implement an init hook, the devhost will reply immediately.

TEST=
runtests -t ddk-lifecycle-test
runtests -t ddk-power-test

The test adds a child device which does not immediately reply to the init hook,
then tries to remove the child device. CompleteChildInit is called to
reply to the init hook. The unbind hook also asserts that init should already be completed.

BUG=38527

Change-Id: I825b4fefc22e6e787b3d61cdda61c0b5b66dd823