[bindc] Add wlan bind library
The library name includes "deprecated" because this is a temporary step
along the migration. The library currently does not define the
variables, instead it only extends variables in the "deprecated."
namespace. This is due to a limitation in the bind bytecode that will be
addressed once all drivers have migrated to using the bind compiler.

Change-Id: I307ea30f5065d0f74543c4406a38c96b9b1e99bb