[bindc] Add video properties to USB bind library
The library name includes "deprecated" because this is a temporary step
along the migration. The library currently does not define the
variables, instead it only extends variables in the "deprecated."
namespace. This is due to a limitation in the bind bytecode that will be
addressed once all drivers have migrated to using the bind compiler.

Change-Id: I8d4f263f8dd4b52194ba7aebba477aab4b52c948