[devcoordinator] stop handing out handle to /hub
On bringup builds pkgfs is never brought up which means the appmgr
component is never run which means the handle to /hub is never serviced,
which means that `ls /` hangs forever when it attempts to look in the
hub handle.

Devcoordinator should stop handing out the /hub handle to children until
the appmgr component doesn't exist on bringup builds.

Bug: 44748
Change-Id: Ifb0f52c5e42f4c87e1b69d3941f5237f2e4b1720