[FDF] Complete rebind and bind fidl calls after init
Rebind and Bind device controller fidl calls are used by the clients to
teardown the entire stack under the device and start binding again.
One of the use cases for these fidl calls is when the partition layout of a
block device changes. These fidl apis will also be useful whenever we want to
unload and reload the driver for a device.As per DDK contract, every driver
adds a device at the end of bind function. If the device needs initialization,
drivers do either of the following:

1.A init hook is implemented and until the hook is complete, the devices
are not visible to the clients.
2.Device is added as INVISIBLE and later made visible after the init.

Device controller BIND/REBIND fidl calls in their current state, respond
once the bind completes. The clients thereafter wait for the devices to
appear by watching the directories.

Since the fidl calls return before the init is complete, the wait for
the devices to appear is going to be indefinite if the init fails.
or if the device is removed as part of initialization.

This change does the following:

1. Makes the rebind/bind fidl calls to complete only after the
initialization is complete.
2. If the device fails initialization, and is removed, these fidl calls
will be replied with ZX_ERR_IO.

This will lead to removal of hacks in storage tests.

Bug:44939

Test: fx run-test driver_host_tests
runtests -t paver-test
runtests -t fvm-test

Change-Id: Iea5fd2ea673133599bd9d8c5becd0b5d031b62d8