[bindc] Implement simple online debugger for bind programs.
This CL implements a simple debugger which can be used to execute the
bind programs of real drivers in the system. It takes the path of a
driver and a device in the system, and executes the bind program of
the driver against the device properties. It outputs whether or not
program would bind, using the same evaluation logic as binding.cc.

Change-Id: Ica17e5e7312793f39f13a25393eff4547ecdc07b