[securemem] Use raw TEE FIDL API
Reland of 95e26eb1c27febfa67f3ad912ac07ce7f60ded8b with additional
testing done. The TA was expecting extra padding on the parameters
being passed over the message buffer, which was previously not
passed, causing errors in the ranges of memory being marked secure.

This change removes the reliance of the securemem driver on an
implementation detail of the TEE Client API and removes the dependency
on tee-client-api.static. Instead, the driver now directly uses the
FIDL interface.

Bug: 36236
Bug: 44664

Test: fx run-test use_vp9_decoder_secure_input_output_test
Test: fx run-test use_h264_decoder_secure_output_test
Test: fx shell runtests -t sysmem-tests
Change-Id: I917311cb77ff329be35761ae11ae9b9a01563f29