[devices][clock] `namespace clock` is invalid C++
The identifier `clock` is a standard C library function.  Such
identifiers are effectively reserved everywhere in C++.  This
matters in practice when a standard C++ library header includes
the standard C library header that declares the function.  Unlike
standard C headers, the C++ standard headers are pretty much
allowed to include any other standard C++ or C header they like.
So when <atomic> starts indirectly including <time.h>, `clock` is
off limits--and the implementation is free to start doing that
any time.  The LLVM libc++ implementation just started doing it.

The uses of `namespace clock` are in a driver and a test.  These
are both places where there is no other code that is exposed to
its external linkage symbols, so there is really no need to put
them into a namespace.

Fixed: 47144
Change-Id: Ic1967d36e97db4c71574393775997cdb00309a2f