[amlogic][sdmmc] Synchronously wait for transfer interrupt
The current design for the interrupt that indicates that an SDIO
transfer has completed involves a separate thread that's waiting for an
interrupt to be signaled. The thread initiating the transfer will then
wait until the interrupt thread has signaled that the transfer is
completed.

This change moves the interrupt wait to the thread doing the transfer.
So instead of waiting for the interrupt thread it just waits directly
for the interrupt. This reduces the impact of context switching costs on
performance. In the broadcom wireless driver this change in combination
with a switch to using DMA sees about a 100% improvement in performance.
However each change individually only makes about a 10% difference.

BUG: 29502
Test: Extensive performance testing using iperf and curl on vim2, astro, sherlock
      aml-sd-emmc-test
Change-Id: Iadcf21d5b02765944bd8f4ee9afc6ccb7b9eff6f