[FDF][Generic-Power-Driver]Maintain device dependencies on power-domain
For power savings, SoC manufacturers group devices of similar functionality and
power them using same power rail usually called power-domain. A power-domain is
further dependent on other power domains based on the power tree. If a
power domain is not being used by any device, the domain can temporarily be
powered down and powered back when needed. It is necessary to maintain the usage
count of the power domain to determine when to power down the domain.
Any voltage modification on a power-domain should be done considering
all devices that are dependent on the domain.

This change adds protocol ops for a device to register for the domain
with the voltages it supports. Any voltage modification request is made
based on supported voltage ranges of all the devices. When a domain is
registered for the very first times, it is enabled by calling into the
power-impl driver. The domain is also registerd with the parent domain,
when dependent count is > 0 and is unregistered from the parent, when
dependent count becomes 0 again.

Test: runtests -t generic-power-driver-test. runtests -t
platform-bus-test. Updated bootloader and paved an astro and ran
device-enumeration-tests on astro. Also verified GI integration is successful.

Change-Id: I4eea17c98b92af601529128168002868f720d1bd