[devices][uart16550] Use hwreg PIO and mock support
This converts the uart16550 driver to take advantage if the new
hwreg::RegisterPio and hwreg::Mock support instead of rolling its own.
As well as factoring out this code, the new mock support also verifies
the relative ordering of PIO reads and writes, which matters to the
correctness of the driver but wasn't tested before.

Also clean up the build rules and don't compile the same file twice
separately for the driver and the test.

Change-Id: I12c45a88e332718323ba6bafbd7e5039c0eec391
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/374194
Commit-Queue: Roland McGrath <mcgrathr@google.com>
Reviewed-by: Tess Eisenberger <teisenbe@google.com>
Testability-Review: Tess Eisenberger <teisenbe@google.com>