[driver_host] Fix leak on fidl methods without reply
We change the logic to return ZX_ERR_ASYNC rather than ZX_OK when a
message was moved to become async. If a method's reply was not invoked
but they didn't move it the reply, then we must decrement the count to
avoid a leak. The specific corner case this addresses is when a fidl
method does not have a reply so correctly returns ZX_OK without calling
the reply method.

Bug: 49138
TESTED: fx test ddk-instance-lifecycle-test
Change-Id: Ibedbc306f2a6080c7a1f6751037e19718ef80a33
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/376317
Reviewed-by: Tess Eisenberger <teisenbe@google.com>
Testability-Review: Tess Eisenberger <teisenbe@google.com>
Commit-Queue: Suraj Malhotra <surajmalhotra@google.com>