[ddk] ShutdownClientConnections after unbind finishes
This patch forcefully shuts down client connections after a driver
replies to Unbind. This is needed to ensure that DdkRelease can
eventually be called if a client does not close its channel.

Additionally; we now wait for a ZX_CHANNEL_PEER_CLOSED
signal even when we are unbound. This is nececssary
to prevent a race where we get a message followed by
an immediate unbind, which previously would result
in the handler being de-registered.

Test: runtests -t ddk-lifecycle-test

Change-Id: Ifc2b2511808d351fedd9f1fb2f67c7cadd7642c6
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/380300
Commit-Queue: Brian Bosak <bbosak@google.com>
Reviewed-by: Tess Eisenberger <teisenbe@google.com>
Testability-Review: Tess Eisenberger <teisenbe@google.com>