[driver_manager] Make devhost logic more self-contained
The driver_host launching logic was mixed in with the rest of the
driver_manager's main logic, though it didn't have any dependenices in
it.  This moves the logic in to the Devhost class to make that more
clear and to allow us to further encapsulate the Devhost's state.

Change-Id: If54e10aa0b79178f6a020711494153f02a62d3e5
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/380556
Commit-Queue: Tess Eisenberger <teisenbe@google.com>
Testability-Review: Tess Eisenberger <teisenbe@google.com>
Reviewed-by: Ruchira Ravoori <ravoorir@google.com>