[FDF]Disable flaking test
Temporarily disabling the test until the actual failure is
investigated.Looks like it is recent, but no changes in resume/suspend
specifically. Since Resume itself is not a feature that is used,
disabling the test is not a drastic measure at this point.

Bug:49398

Change-Id: Ib81803ea8ef39c91ebc2d8160f7b2e2615f3fd8d
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/384256
Reviewed-by: Puneetha Ramachandra <puneetha@google.com>
Testability-Review: Puneetha Ramachandra <puneetha@google.com>
Commit-Queue: Ruchira Ravoori <ravoorir@google.com>