[sysmem] Add API for setting BufferCollection name
We want participants that don't receive VMOs to be able to name them,
because they might have better high-level ideas of what the VMOs are
for. So add a SetName method to buffer collection to allow any
participant to set the VMO name. This also allows passing through the
name to the Allocator, so it can print out buffer names when it runs out
of space.

Add support to magma and the amlogic video decoder drivers for setting
this property, so we can take advantage of the new feature immediately.

Change-Id: Ic36652529e7adb7261110f80b148a95701d21b13
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/387389
Reviewed-by: Dustin Green <dustingreen@google.com>
Testability-Review: John Bauman <jbauman@google.com>
Commit-Queue: John Bauman <jbauman@google.com>
API-Review: Dale Sather <dalesat@google.com>