[ddk][devmgr] Move away from custom fbl container traits.
Update code to stop using custom fbl container traits in various
places in the devmgr.

This refactor removes a multiple uses of custom container traits from
the system which helps to prepare for an up-and-coming refactor of the
fbl:: container template arguments which will require special handling
of users of custom node traits.  The fewer of these that we have, the
easier the refactor will be.

Test: existing unit tests
Change-Id: I75bfa918097e56a1e184fa41efeba9bb6ad6ba4c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/386824
Commit-Queue: John Grossman <johngro@google.com>
Reviewed-by: George Kulakowski <kulakowski@google.com>
Reviewed-by: Suraj Malhotra <surajmalhotra@google.com>
Testability-Review: Suraj Malhotra <surajmalhotra@google.com>