[ddk][devhost] Move away from custom fbl container traits.
Update code to stop using custom fbl container traits in order to
allow zx_device objects to exist in multiple containers
simultaneously.

This refactor removes a use of custom container traits from the system
which helps to prepare for an up-and-coming refactor of the fbl::
container template arguments which will require special handling of
users of custom node traits.  The fewer of these that we have, the
easier the refactor will be.

Test: existing unit tests

Change-Id: I53faa6471aa5feafa8ae4708d40f8cc5fe65b484
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/386823
Commit-Queue: John Grossman <johngro@google.com>
Reviewed-by: Nick Maniscalco <maniscalco@google.com>
Reviewed-by: George Kulakowski <kulakowski@google.com>
Testability-Review: Nick Maniscalco <maniscalco@google.com>