[sysmem] Add index to vmo names
This makes it easier for clients to identify which VMO is leaking, if
only a subset of their VMOs have problems. Also add the koid on to the
logging so that we can correlate sysmem's information with that from
memgraph -v.

Change-Id: I4aac6431b474a201ba4b0a448b4b567750d332eb
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/389217
Testability-Review: John Bauman <jbauman@google.com>
Commit-Queue: John Bauman <jbauman@google.com>
Reviewed-by: Dustin Green <dustingreen@google.com>