[virtio][socket] Move away from custom fbl container traits.
Update code to stop using custom fbl container traits in order to
allow SocketDevice::Connection objects to exist in multiple containers
simultaneously.

This refactor removes a use of custom container traits from the system
which helps to prepare for an up-and-coming refactor of the fbl::
container template arguments which will require special handling of
users of custom node traits.  The fewer of these that we have, the
easier the refactor will be.

Test: existing unit tests
Change-Id: Iefcf4d52949045a28a8c49d80b3cd9bdaa50abb5
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/386822
Commit-Queue: John Grossman <johngro@google.com>
Reviewed-by: Ruchira Ravoori <ravoorir@google.com>
Reviewed-by: Christopher Anderson <cja@google.com>
Testability-Review: Christopher Anderson <cja@google.com>